+++ lesson-5/config/wechat_file.txt
@@ -1,1 +1,1 @@
-{"access_token":"ZAeBzph26GJXgUcwUrLH9GNs68qXWFgplkPLqm2q6KDzcjYWl8BxtQc6x8oYhXcK_aKj9mxy6VsesPOaLaulQXyVk-GAN20Klwo2GHW7ksDpGLBTT2m8Kz4OCK8swneHSXCbAJAMRW","expires_in":1463129907787}
+{"access_token":"CP2kFsvG_AZUMDclwSLLo1Pn3A1KQTeVvBOFGTPp-xG3PmXmwSt-XpuPnBjMHzYDj5bb4Cx0wc1zzNLUxDGfd6F5WEGFVQZqPT7KMrVCaZlW0NOByZv7PD1JGypxb3nfHZPbAJAYQG","expires_in":1463137429841}
